refactor(Flow cell encryption) (#3358) (patch)
# Description
Refactor the flow cell encryption api.

##  Added
## Changed
Renamed FlowCellEncryptionAPI to IlluminaRunEncryptionService

Moved IlluminaRunEncryptionService` to cg/services/illumina_services/backup_services/encryption_services

Move encryption fixtures to a dedicated fixute plugin file

Moved and refactored tests of the ``IlluminaRunEncryptionService

Renamend cli command cg encrypt flow-cells to cg encrypt illumina-runs

Fixed
