+++ cg/meta/encryption/sbatch.py
@@ -1,24 +1,24 @@
 """Hold encryption sbatch templates in variables."""
 
-FLOW_CELL_ENCRYPT_ERROR = """
+ILLUMINA_RUN_ENCRYPT_ERROR = """
 if [[ -e {pending_file_path} ]]
 then
     rm -f {pending_file_path}
 fi
 """
 
-FLOW_CELL_ENCRYPT_COMMANDS = """
+ILLUMINA_RUN_ENCRYPT_COMMANDS = """
 {symmetric_passphrase}
 
 {asymmetrically_encrypt_passphrase}
 
 {make_tmp_encrypt_dir}
 
-{copy_flow_cell_dir_to_tmp}
+{copy_run_dir_to_tmp}
 
-{tar_encrypt_tmp_dir} | {parallel_gzip} | {tee} | {flow_cell_symmetric_encryption}
+{tar_encrypt_tmp_dir} | {parallel_gzip} | {tee} | {run_symmetric_encryption}
 
-{flow_cell_symmetric_decryption} | {md5sum}
+{run_symmetric_decryption} | {md5sum}
 
 {diff}
