+++ cg/services/pdc_service/pdc_service.py
@@ -11,10 +11,12 @@ from cg.constants.process import EXIT_WARNING
 from cg.exc import (
     DsmcAlreadyRunningError,
     FlowCellAlreadyBackedUpError,
-    FlowCellEncryptionError,
+    IlluminaRunEncryptionError,
     PdcError,
 )
-from cg.meta.encryption.encryption import FlowCellEncryptionAPI
+from cg.services.illumina_services.backup_services.encrypt_service import (
+    IlluminaRunEncryptionService,
+)
 from cg.store.models import Flowcell
 from cg.store.store import Store
 from cg.utils import Process
