+++ tests/cli/backup/test_backup_command.py
@@ -1,13 +1,17 @@
 import logging
 from pathlib import Path
 
+import pytest
 from click.testing import CliRunner
 from psutil import Process
 
-from cg.cli.backup import backup_flow_cells, encrypt_flow_cells, fetch_flow_cell
+from cg.cli.backup import backup_flow_cells, encrypt_illumina_runs, fetch_flow_cell
 from cg.constants import EXIT_SUCCESS, FileExtensions, FlowCellStatus
+from cg.exc import IlluminaRunEncryptionError
 from cg.models.cg_config import CGConfig
 from cg.models.run_devices.illumina_run_directory_data import IlluminaRunDirectoryData
+from cg.store.models import IlluminaSequencingRun
+from cg.store.store import Store
 from tests.store_helpers import StoreHelpers
 
 
@@ -117,141 +121,148 @@ def test_backup_flow_cells_when_encryption_is_not_completed(
     assert f"Flow cell: {flow_cell_name} encryption process is not complete" in caplog.text
 
 
-def test_encrypt_flow_cells(
+def test_encrypt_illumina_runs(
     cli_runner: CliRunner, cg_context: CGConfig, caplog, sbatch_job_number: str
 ):
     """Test encrypt flow cell in dry run mode."""
     caplog.set_level(logging.INFO)
 
-    # GIVEN a flow cells directory
+    # GIVEN an illumina runs directory
 
-    # WHEN encrypting flow cells in dry run mode
-    result = cli_runner.invoke(encrypt_flow_cells, ["--dry-run"], obj=cg_context)
+    # WHEN encrypting run in dry run mode
+    result = cli_runner.invoke(encrypt_illumina_runs, ["--dry-run"], obj=cg_context)
 
     # THEN exits without any errors
     assert result.exit_code == EXIT_SUCCESS
 
-    # THEN communicate flow cell encryption is submitted
-    assert f"Flow cell encryption running as job {sbatch_job_number}" in caplog.text
+    # THEN communicate encryption job is submitted
+    assert f"Run encryption running as job {sbatch_job_number}" in caplog.text
 
 
-def test_encrypt_flow_cells_when_already_backed_up(
+def test_encrypt_illumina_runs_when_already_backed_up(
     cli_runner: CliRunner,
     cg_context: CGConfig,
     caplog,
-    flow_cell_name: str,
+    novaseq_x_flow_cell_id: str,
+    store_with_illumina_sequencing_data: Store,
     helpers: StoreHelpers,
 ):
-    """Test encrypt flow cell in dry run mode when there is already a back-up."""
+    """Test encrypt illumina run in dry run mode when there is already a back-up."""
     caplog.set_level(logging.DEBUG)
 
-    # Given a flow cell with a back-up
-    helpers.add_flow_cell(
-        store=cg_context.status_db, flow_cell_name=flow_cell_name, has_backup=True
+    # GIVEN a sequencing run with a back-up
+    cg_context.status_db_ = store_with_illumina_sequencing_data
+    sequencing_run: IlluminaSequencingRun = (
+        store_with_illumina_sequencing_data.get_illumina_sequencing_run_by_device_internal_id(
+            novaseq_x_flow_cell_id
+        )
     )
+    sequencing_run.has_backup = True
 
-    # GIVEN a flow cells directory
+    # GIVEN a sequencing runs directory
 
-    # WHEN encrypting flow cells in dry run mode
-    result = cli_runner.invoke(encrypt_flow_cells, ["--dry-run"], obj=cg_context)
+    # WHEN encrypting runs in dry run mode
+    result = cli_runner.invoke(encrypt_illumina_runs, ["--dry-run"], obj=cg_context)
 
     # THEN exits without any errors
     assert result.exit_code == EXIT_SUCCESS
 
-    # THEN communicate flow cell is already backed-up
-    assert f"Flow cell: {flow_cell_name} is already backed-up" in caplog.text
+    # THEN communicate run is already backed-up
+    assert f"Run: {novaseq_x_flow_cell_id} is already backed-up" in caplog.text
 
 
-def test_encrypt_flow_cells_when_sequencing_not_done(
+def test_encrypt_illumina_run_when_sequencing_not_done(
     cli_runner: CliRunner,
     cg_context: CGConfig,
     caplog,
     mocker,
-    flow_cell_name: str,
+    novaseq_x_flow_cell_id: str,
 ):
-    """Test encrypt flow cell in dry run mode when sequencing is not done."""
+    """Test encrypt illumina runs in dry run mode when sequencing is not done."""
     caplog.set_level(logging.DEBUG)
 
     # GIVEN flow cells that are being sequenced
     mocker.patch.object(IlluminaRunDirectoryData, "is_sequencing_run_ready")
     IlluminaRunDirectoryData.is_sequencing_run_ready.return_value = False
 
-    # GIVEN a flow cells directory
+    # GIVEN a sequencing runs directory
 
-    # WHEN encrypting flow cells in dry run mode
-    result = cli_runner.invoke(encrypt_flow_cells, ["--dry-run"], obj=cg_context)
+    # WHEN encrypting runs in dry run mode
+    result = cli_runner.invoke(encrypt_illumina_runs, ["--dry-run"], obj=cg_context)
 
     # THEN exits without any errors
     assert result.exit_code == EXIT_SUCCESS
 
-    # THEN communicate flow cell is not ready
-    assert f"Flow cell: {flow_cell_name} is not ready" in caplog.text
+    # THEN communicate the run is not ready
+    assert f"Run: {novaseq_x_flow_cell_id} is not ready" in caplog.text
 
 
-def test_encrypt_flow_cell_when_encryption_already_started(
+def test_encrypt_illumina_run_when_encryption_already_started(
     cli_runner: CliRunner,
-    cg_context: CGConfig,
+    encryption_context: CGConfig,
     caplog,
     pdc_archiving_dir: Path,
-    flow_cell_name: str,
-    flow_cell_full_name: str,
+    novaseq_x_flow_cell_id: str,
+    novaseq_x_flow_cell_full_name: str,
+    store_with_illumina_sequencing_data: Store,
+    tmp_path: Path,
     mocker,
 ):
-    """Test encrypt flow cell in dry run mode when pending file exists"""
+    """Test encrypt illumina runs in dry run mode when pending file exists"""
     caplog.set_level(logging.DEBUG)
 
-    # GIVEN flow cells that are ready
+    # GIVEN illumina runs that are ready
     mocker.patch.object(IlluminaRunDirectoryData, "is_sequencing_run_ready")
     IlluminaRunDirectoryData.is_sequencing_run_ready.return_value = True
 
     # GIVEN a pending flag file
-    flow_cells_encrypt_dir = Path(cg_context.encryption.encryption_dir, flow_cell_full_name)
-    flow_cells_encrypt_dir.mkdir(parents=True, exist_ok=True)
-    Path(flow_cells_encrypt_dir, flow_cell_name).with_suffix(FileExtensions.PENDING).touch()
-
-    # GIVEN a flow cells directory
+    Path(
+        encryption_context.encryption.encryption_dir,
+        novaseq_x_flow_cell_full_name,
+        novaseq_x_flow_cell_id,
+    ).with_suffix(FileExtensions.PENDING).touch()
 
     # WHEN encrypting flow cells in dry run mode
-    result = cli_runner.invoke(encrypt_flow_cells, ["--dry-run"], obj=cg_context)
+    result = cli_runner.invoke(encrypt_illumina_runs, ["--dry-run"], obj=encryption_context)
 
     # THEN exits without any errors
     assert result.exit_code == EXIT_SUCCESS
 
     # THEN communicate flow cell encryption already started
-    assert f"Encryption already started for flow cell: {flow_cell_name}" in caplog.text
+    assert f"Encryption already started for run: {novaseq_x_flow_cell_id}" in caplog.text
 
 
 def test_encrypt_flow_cell_when_encryption_already_completed(
     cli_runner: CliRunner,
-    cg_context: CGConfig,
+    encryption_context: CGConfig,
+    novaseq_x_flow_cell_full_name: str,
+    novaseq_x_flow_cell_id: str,
     caplog,
     pdc_archiving_dir: Path,
-    flow_cell_name: str,
-    flow_cell_full_name: str,
     mocker,
 ):
-    """Test encrypt flow cell in dry run mode when completed file exists"""
+    """Test encrypt illumina runs in dry run mode when completed file exists"""
     caplog.set_level(logging.DEBUG)
 
-    # GIVEN flow cells that are ready
+    # GIVEN illumina runs that are ready
     mocker.patch.object(IlluminaRunDirectoryData, "is_sequencing_run_ready")
     IlluminaRunDirectoryData.is_sequencing_run_ready.return_value = True
 
     # GIVEN a complete flag file
-    flow_cells_encrypt_dir = Path(cg_context.encryption.encryption_dir, flow_cell_full_name)
-    flow_cells_encrypt_dir.mkdir(parents=True, exist_ok=True)
-    Path(flow_cells_encrypt_dir, flow_cell_name).with_suffix(FileExtensions.COMPLETE).touch()
-
-    # GIVEN a flow cells directory
+    Path(
+        encryption_context.encryption.encryption_dir,
+        novaseq_x_flow_cell_full_name,
+        novaseq_x_flow_cell_id,
+    ).with_suffix(FileExtensions.COMPLETE).touch()
 
     # WHEN encrypting flow cells in dry run mode
-    result = cli_runner.invoke(encrypt_flow_cells, ["--dry-run"], obj=cg_context)
+    result = cli_runner.invoke(encrypt_illumina_runs, ["--dry-run"], obj=encryption_context)
 
     # THEN exits without any errors
     assert result.exit_code == EXIT_SUCCESS
 
     # THEN communicate flow cell encryption already completed
-    assert f"Encryption already completed for flow cell: {flow_cell_name}" in caplog.text
+    assert f"Encryption already completed for run: {novaseq_x_flow_cell_id}" in caplog.text
 
 
 def test_run_fetch_flow_cell_dry_run_no_flow_cell_specified(
