+++ tests/meta/encryption/test_encryption.py
@@ -2,19 +2,14 @@
 
 import logging
 import pathlib
-import shutil
 from pathlib import Path
 
 import mock
-import pytest
 
-from cg.exc import FlowCellEncryptionError, FlowCellError
 from cg.meta.encryption.encryption import (
     EncryptionAPI,
-    FlowCellEncryptionAPI,
     SpringEncryptionAPI,
 )
-from cg.models.run_devices.illumina_run_directory_data import IlluminaRunDirectoryData
 
 
 @mock.patch("cg.utils.Process")
@@ -301,193 +296,3 @@ def test_cleanup_no_files(
     assert "No encrypted spring file to clean up, continuing cleanup" in caplog.text
     assert "No encrypted key file to clean up, continuing cleanup" in caplog.text
     assert "No existing key file to clean up, cleanup process completed" in caplog.text
-
-
-def test_flow_cell_encryption_api(flow_cell_encryption_api: FlowCellEncryptionAPI):
-    """Tests instantiating flow cell encryption API."""
-    # GIVEN a FlowCellEncryptionAPI
-
-    # WHEN instantiating the API
-
-    # THEN return a FlowCellEncryptionAPI object
-    assert isinstance(flow_cell_encryption_api, FlowCellEncryptionAPI)
-
-
-def test_get_flow_cell_symmetric_encryption_command(
-    flow_cell_encryption_api: FlowCellEncryptionAPI,
-    output_file_path: Path,
-    temporary_passphrase: Path,
-):
-    # GIVEN a FlowCellEncryptionAPI
-
-    # WHEN getting the command
-    command: str = flow_cell_encryption_api.get_flow_cell_symmetric_encryption_command(
-        output_file=output_file_path, passphrase_file_path=temporary_passphrase
-    )
-
-    # THEN return a string
-    assert isinstance(command, str)
-
-
-def test_get_flow_cell_symmetric_decryption_command(
-    flow_cell_encryption_api: FlowCellEncryptionAPI,
-    input_file_path: Path,
-    temporary_passphrase: Path,
-):
-    # GIVEN a FlowCellEncryptionAPI
-
-    # WHEN getting the command
-    command: str = flow_cell_encryption_api.get_flow_cell_symmetric_decryption_command(
-        input_file=input_file_path, passphrase_file_path=temporary_passphrase
-    )
-
-    # THEN return a string
-    assert isinstance(command, str)
-
-
-def test_create_pending_file(
-    flow_cell_encryption_api: FlowCellEncryptionAPI,
-):
-    # GIVEN a FlowCellEncryptionAPI
-
-    # WHEN checking if encryption is possible
-    flow_cell_encryption_api.flow_cell_encryption_dir.mkdir(parents=True)
-    flow_cell_encryption_api.dry_run = False
-    flow_cell_encryption_api.create_pending_file(
-        pending_path=flow_cell_encryption_api.pending_file_path
-    )
-
-    # THEN pending file should exist
-    assert flow_cell_encryption_api.pending_file_path.exists()
-
-    # Clean-up
-    shutil.rmtree(flow_cell_encryption_api.flow_cell_encryption_dir)
-
-
-def test_create_pending_file_when_dry_run(
-    flow_cell_encryption_api: FlowCellEncryptionAPI,
-):
-    # GIVEN a FlowCellEncryptionAPI
-
-    # WHEN checking if encryption is possible
-    flow_cell_encryption_api.create_pending_file(
-        pending_path=flow_cell_encryption_api.pending_file_path
-    )
-
-    # THEN pending file should not exist
-    assert not flow_cell_encryption_api.pending_file_path.exists()
-
-
-def test_is_encryption_possible(flow_cell_encryption_api: FlowCellEncryptionAPI, mocker):
-    # GIVEN a FlowCellEncryptionAPI
-
-    # GIVEN that sequencing is ready
-    mocker.patch.object(IlluminaRunDirectoryData, "is_sequencing_run_ready")
-    IlluminaRunDirectoryData.is_sequencing_run_ready.return_value = True
-
-    # WHEN checking if encryption is possible
-    is_possible: bool = flow_cell_encryption_api.is_encryption_possible()
-
-    # THEN return True
-    assert is_possible
-
-
-def test_is_encryption_possible_when_sequencing_not_ready(
-    caplog, flow_cell_encryption_api: FlowCellEncryptionAPI, flow_cell_name: str, mocker
-):
-    caplog.set_level(logging.ERROR)
-
-    # GIVEN a FlowCellEncryptionAPI
-
-    # GIVEN that sequencing is not ready
-    mocker.patch.object(IlluminaRunDirectoryData, "is_sequencing_run_ready")
-    IlluminaRunDirectoryData.is_sequencing_run_ready.return_value = False
-
-    # WHEN checking if encryption is possible
-    with pytest.raises(FlowCellError):
-        flow_cell_encryption_api.is_encryption_possible()
-
-        # THEN error should be raised
-        assert f"Flow cell: {flow_cell_name} is not ready" in caplog.text
-
-
-def test_is_encryption_possible_when_encryption_is_completed(
-    caplog, flow_cell_encryption_api: FlowCellEncryptionAPI, flow_cell_name: str, mocker
-):
-    # GIVEN a FlowCellEncryptionAPI
-
-    # GIVEN that sequencing is ready
-    mocker.patch.object(IlluminaRunDirectoryData, "is_sequencing_run_ready")
-    IlluminaRunDirectoryData.is_sequencing_run_ready.return_value = True
-
-    # GIVEN that encryption is completed
-    flow_cell_encryption_api.flow_cell_encryption_dir.mkdir(parents=True)
-    flow_cell_encryption_api.complete_file_path.touch()
-
-    # WHEN checking if encryption is possible
-    with pytest.raises(FlowCellEncryptionError):
-        flow_cell_encryption_api.is_encryption_possible()
-
-        # THEN error should be raised
-        assert f"Encryption already completed for flow cell: {flow_cell_name}" in caplog.text
-
-    # Clean-up
-    shutil.rmtree(flow_cell_encryption_api.flow_cell_encryption_dir)
-
-
-def test_is_encryption_possible_when_encryption_is_pending(
-    caplog, flow_cell_encryption_api: FlowCellEncryptionAPI, flow_cell_name: str, mocker
-):
-    # GIVEN a FlowCellEncryptionAPI
-
-    # GIVEN that sequencing is ready
-    mocker.patch.object(IlluminaRunDirectoryData, "is_sequencing_run_ready")
-    IlluminaRunDirectoryData.is_sequencing_run_ready.return_value = True
-
-    # GIVEN that encryption is pending
-    flow_cell_encryption_api.flow_cell_encryption_dir.mkdir(parents=True)
-    flow_cell_encryption_api.pending_file_path.touch()
-
-    # WHEN checking if encryption is possible
-    with pytest.raises(FlowCellEncryptionError):
-        flow_cell_encryption_api.is_encryption_possible()
-
-        # THEN error should be raised
-        assert f"Encryption already started for flow cell: {flow_cell_name}" in caplog.text
-
-    # Clean-up
-    shutil.rmtree(flow_cell_encryption_api.flow_cell_encryption_dir)
-
-
-def test_encrypt_flow_cell(
-    caplog, flow_cell_encryption_api: FlowCellEncryptionAPI, mocker, sbatch_job_number: int
-):
-    caplog.set_level(logging.INFO)
-    # GIVEN a FlowCellEncryptionAPI
-
-    # GIVEN that sequencing is ready
-    mocker.patch.object(IlluminaRunDirectoryData, "is_sequencing_run_ready")
-    IlluminaRunDirectoryData.is_sequencing_run_ready.return_value = True
-
-    # WHEN encrypting flow cell
-    flow_cell_encryption_api.encrypt_flow_cell()
-
-    # THEN sbatch should be submitted
-    assert f"Flow cell encryption running as job {sbatch_job_number}" in caplog.text
-
-
-def test_start_encryption(
-    caplog, flow_cell_encryption_api: FlowCellEncryptionAPI, mocker, sbatch_job_number: int
-):
-    caplog.set_level(logging.INFO)
-    # GIVEN a FlowCellEncryptionAPI
-
-    # GIVEN that sequencing is ready
-    mocker.patch.object(IlluminaRunDirectoryData, "is_sequencing_run_ready")
-    IlluminaRunDirectoryData.is_sequencing_run_ready.return_value = True
-
-    # WHEN trying to start encrypting flow cell
-    flow_cell_encryption_api.start_encryption()
-
-    # THEN sbatch should be submitted
-    assert f"Flow cell encryption running as job {sbatch_job_number}" in caplog.text
